unroll and vectorization: add gauss elimination
This directory contains source code of a linear system resolution usign
Gauß elimination then back substitution.
It is one of the general purpose code we used to test different aspects
of MAQAO.
It exists in the three languages supported by MAQAO.
We wanted to look at the advices given by the tool when a different language is used.

This directory also contains assembly code of the C version compile by
gcc and disassemble by objdump.
We used those files to verify the disassembly operated by MAQAO's module
madras.