Add MPI reductions and time folder
The aim of those code is to verify general information on pure MPI
program.
It helps confirm the time given by MAQAO and is purposed to be used with
different MPI runtime.

The two source code file contains a reduction of a vector several times.
The simplicity of the operation combined with a well chosen size of
array is supposed to lead to efficiencies over one on certain run.

A README.md has been added to give more details.